#15 Refactored Index Resolver to use the previously refactored overload resolution support
