+++ IronVelocity/Reflection/IOverloadResolver.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Dynamic;
+using System.Linq.Expressions;
 using System.Reflection;
 
 namespace IronVelocity.Reflection
@@ -7,6 +9,6 @@ namespace IronVelocity.Reflection
     public interface IOverloadResolver
     {
         FunctionMemberData<T> Resolve<T>(IEnumerable<FunctionMemberData<T>> candidates, Type[] args);
-        bool IsParameterArrayArgument(ParameterInfo parameter);
+        Expression[] CreateParameterExpressions(ParameterInfo[] parameters, DynamicMetaObject[] args);
     }
 }
