+++ IronVelocity/Reflection/OverloadResolution.cs
@@ -1,6 +1,9 @@
-﻿using System;
+﻿using IronVelocity.Compilation;
+using System;
 using System.Collections.Generic;
+using System.Dynamic;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Reflection;
 
 namespace IronVelocity.Reflection
@@ -32,6 +35,56 @@ namespace IronVelocity.Reflection
             //members with respect to the given argument list, provided that each function member is compared to
             //all other function members using the rules in §7.5.3.2.
             return GetBestFunctionMember(aplicableCandidateFunctionMembers);
+        }
+
+        public Expression[] CreateParameterExpressions(ParameterInfo[] parameters, DynamicMetaObject[] args)
+        {
+            if (parameters == null)
+                throw new ArgumentNullException(nameof(parameters));
+
+            if (args == null)
+                throw new ArgumentNullException(nameof(args));
+
+            var lastParameter = parameters.LastOrDefault();
+            bool hasParamsArray = IsParameterArrayArgument(lastParameter);
+
+            int trivialParams = hasParamsArray
+                ? parameters.Length - 1
+                : parameters.Length;
+
+            var argTypeArray = args
+                .Select(x => x.Value == null ? null : x.LimitType)
+                .ToArray();
+
+
+            var argExpressions = new Expression[parameters.Length];
+            for (int i = 0; i < trivialParams; i++)
+            {
+                var parameter = parameters[i];
+                argExpressions[i] = VelocityExpressions.ConvertParameterIfNeeded(args[i], parameter);
+            }
+            if (hasParamsArray)
+            {
+                int lastIndex = argExpressions.Length - 1;
+                //Check if the array has been explicitly passed, rather than as individual elements
+                if (args.Length == parameters.Length
+                    && _argumentConverter.CanBeConverted(argTypeArray.Last(), lastParameter.ParameterType)
+                    && argTypeArray.Last() != null)
+                {
+                    argExpressions[lastIndex] = VelocityExpressions.ConvertParameterIfNeeded(args[lastIndex], lastParameter);
+                }
+                else
+                {
+                    var elementType = lastParameter.ParameterType.GetElementType();
+                    argExpressions[lastIndex] = Expression.NewArrayInit(
+                        elementType,
+                        args.Skip(lastIndex)
+                            .Select(x => VelocityExpressions.ConvertIfNeeded(x, elementType))
+                        );
+                }
+            }
+
+            return argExpressions;
         }
 
         private FunctionMemberData<T> GetBestFunctionMember<T>(IReadOnlyCollection<FunctionMemberData<T>> candidates)
